Move replay builder internally
