+++ src/index.ts
@@ -3,3 +3,5 @@ export { default as FPSClock } from "./utils/FPSClock"
 export { ReplayData } from "./models/ReplayData"
 export { ReplayMetadata } from "./models/ReplayMetadata"
 export { GameManager } from "./managers/GameManager"
+export { loadReplay } from "./viewer/clients/loadReplay"
+export { loadBuilderFromReplay } from "./viewer/clients/loadBuilderFromReplay"
